Fix for edge case when gif data is invalid
Sometimes GIF data may be invalid due to incorrect encoding, corrupted transmission etc. This causes problems when attempting to call `drawImage()` to the canvas context. In most cases, the frame data will be `undefined` so we must check for this before attempting `drawImage()` just in case. There’s probably a more elegant way to handle this, but it seems to work for my use-case.